Refactor MainViewModel: extract commands into partial classes
Split large InitializeCommands() method into 6 partial class files organized
by functional area for improved maintainability and code navigation:

- MainViewModel.Commands.Navigation.cs: panel toggles, view settings, camera
- MainViewModel.Commands.Simulator.cs: simulator controls and movement
- MainViewModel.Commands.Configuration.cs: config dialogs, profiles, DataIO
- MainViewModel.Commands.Fields.cs: field selection, creation, imports
- MainViewModel.Commands.Boundary.cs: boundary recording, headland management
- MainViewModel.Commands.Track.cs: tracks, AB lines, guidance, sections

Reduces MainViewModel.cs from ~8200 to ~5820 lines while preserving all
functionality and XAML bindings.
