+++ Shared/AgValoniaGPS.ViewModels/MainViewModel.Commands.Configuration.cs
@@ -1,0 +1,178 @@
+using System.Windows.Input;
+using CommunityToolkit.Mvvm.Input;
+using AgValoniaGPS.Models.State;
+
+namespace AgValoniaGPS.ViewModels;
+
+/// <summary>
+/// Configuration commands - vehicle profiles, configuration dialogs, DataIO
+/// </summary>
+public partial class MainViewModel
+{
+    // Configuration Dialog Commands
+    public ICommand? ShowConfigurationDialogCommand { get; private set; }
+    public ICommand? CancelConfigurationDialogCommand { get; private set; }
+
+    // Profile Selection Commands
+    public ICommand? ShowLoadProfileDialogCommand { get; private set; }
+    public ICommand? ShowNewProfileDialogCommand { get; private set; }
+    public ICommand? LoadSelectedProfileCommand { get; private set; }
+    public ICommand? CancelProfileSelectionCommand { get; private set; }
+
+    // DataIO Dialog Commands
+    public ICommand? ShowDataIODialogCommand { get; private set; }
+    public ICommand? CloseDataIODialogCommand { get; private set; }
+
+    // AgShare Dialog Commands
+    public ICommand? ShowAgShareDownloadDialogCommand { get; private set; }
+    public ICommand? CancelAgShareDownloadDialogCommand { get; private set; }
+    public ICommand? ShowAgShareUploadDialogCommand { get; private set; }
+    public ICommand? CancelAgShareUploadDialogCommand { get; private set; }
+    public ICommand? ShowAgShareSettingsDialogCommand { get; private set; }
+    public ICommand? CancelAgShareSettingsDialogCommand { get; private set; }
+    public ICommand? ConfirmAgShareSettingsDialogCommand { get; private set; }
+
+    // Numeric Input Dialog Commands
+    public ICommand? CancelNumericInputDialogCommand { get; private set; }
+    public ICommand? ConfirmNumericInputDialogCommand { get; private set; }
+
+    private void InitializeConfigurationCommands()
+    {
+        // Configuration Dialog Commands
+        ShowConfigurationDialogCommand = new RelayCommand(() =>
+        {
+            ConfigurationViewModel = new ConfigurationViewModel(_configurationService);
+            ConfigurationViewModel.CloseRequested += (s, e) =>
+            {
+                ConfigurationViewModel.IsDialogVisible = false;
+            };
+            ConfigurationViewModel.IsDialogVisible = true;
+        });
+
+        CancelConfigurationDialogCommand = new RelayCommand(() =>
+        {
+            if (ConfigurationViewModel != null)
+                ConfigurationViewModel.IsDialogVisible = false;
+        });
+
+        // Profile Selection Commands
+        ShowLoadProfileDialogCommand = new RelayCommand(() =>
+        {
+            // Refresh available profiles
+            AvailableProfiles.Clear();
+            foreach (var profile in _configurationService.GetAvailableProfiles())
+            {
+                AvailableProfiles.Add(profile);
+            }
+            SelectedProfile = _configurationService.Store.ActiveProfileName;
+            IsProfileSelectionVisible = true;
+        });
+
+        LoadSelectedProfileCommand = new RelayCommand(() =>
+        {
+            if (!string.IsNullOrEmpty(SelectedProfile))
+            {
+                _configurationService.LoadProfile(SelectedProfile);
+                _settingsService.Settings.LastUsedVehicleProfile = SelectedProfile;
+                _settingsService.Save();
+                OnPropertyChanged(nameof(CurrentProfileName));
+            }
+            IsProfileSelectionVisible = false;
+        });
+
+        CancelProfileSelectionCommand = new RelayCommand(() =>
+        {
+            IsProfileSelectionVisible = false;
+        });
+
+        ShowNewProfileDialogCommand = new RelayCommand(() =>
+        {
+            // Generate a unique profile name
+            var baseName = "New Profile";
+            var profileName = baseName;
+            var counter = 1;
+            var existingProfiles = _configurationService.GetAvailableProfiles();
+            while (existingProfiles.Contains(profileName))
+            {
+                profileName = $"{baseName} {counter++}";
+            }
+
+            _configurationService.CreateProfile(profileName);
+            _settingsService.Settings.LastUsedVehicleProfile = profileName;
+            _settingsService.Save();
+            OnPropertyChanged(nameof(CurrentProfileName));
+        });
+
+        // DataIO Dialog Commands
+        ShowDataIODialogCommand = new RelayCommand(() =>
+        {
+            State.UI.ShowDialog(DialogType.DataIO);
+        });
+
+        CloseDataIODialogCommand = new RelayCommand(CloseDataIODialog);
+
+        // AgShare Dialog Commands
+        ShowAgShareDownloadDialogCommand = new RelayCommand(() =>
+        {
+            State.UI.ShowDialog(DialogType.AgShareDownload);
+        });
+
+        CancelAgShareDownloadDialogCommand = new RelayCommand(() =>
+        {
+            State.UI.CloseDialog();
+        });
+
+        ShowAgShareUploadDialogCommand = new RelayCommand(() =>
+        {
+            State.UI.ShowDialog(DialogType.AgShareUpload);
+        });
+
+        CancelAgShareUploadDialogCommand = new RelayCommand(() =>
+        {
+            State.UI.CloseDialog();
+        });
+
+        ShowAgShareSettingsDialogCommand = new RelayCommand(() =>
+        {
+            // Load current settings from storage
+            AgShareSettingsServerUrl = _settingsService.Settings.AgShareServer;
+            AgShareSettingsApiKey = _settingsService.Settings.AgShareApiKey;
+            AgShareSettingsEnabled = _settingsService.Settings.AgShareEnabled;
+            State.UI.ShowDialog(DialogType.AgShareSettings);
+        });
+
+        CancelAgShareSettingsDialogCommand = new RelayCommand(() =>
+        {
+            State.UI.CloseDialog();
+        });
+
+        ConfirmAgShareSettingsDialogCommand = new RelayCommand(() =>
+        {
+            // Save settings to storage
+            _settingsService.Settings.AgShareServer = AgShareSettingsServerUrl;
+            _settingsService.Settings.AgShareApiKey = AgShareSettingsApiKey;
+            _settingsService.Settings.AgShareEnabled = AgShareSettingsEnabled;
+            _settingsService.Save();
+
+            State.UI.CloseDialog();
+            StatusMessage = "AgShare settings saved";
+        });
+
+        // Numeric Input Dialog Commands
+        CancelNumericInputDialogCommand = new RelayCommand(() =>
+        {
+            State.UI.CloseDialog();
+            _numericInputDialogCallback = null;
+        });
+
+        ConfirmNumericInputDialogCommand = new RelayCommand(() =>
+        {
+            if (NumericInputDialogValue.HasValue && _numericInputDialogCallback != null)
+            {
+                _numericInputDialogCallback((double)NumericInputDialogValue.Value);
+            }
+            State.UI.CloseDialog();
+            _numericInputDialogCallback = null;
+        });
+    }
+}
