+++ Shared/AgValoniaGPS.ViewModels/MainViewModel.Commands.Navigation.cs
@@ -1,0 +1,155 @@
+using System.Windows.Input;
+using CommunityToolkit.Mvvm.Input;
+
+namespace AgValoniaGPS.ViewModels;
+
+/// <summary>
+/// Navigation panel commands - panel toggles, view settings, camera controls
+/// </summary>
+public partial class MainViewModel
+{
+    // Panel Toggle Commands
+    public ICommand? ToggleViewSettingsPanelCommand { get; private set; }
+    public ICommand? ToggleFileMenuPanelCommand { get; private set; }
+    public ICommand? ToggleToolsPanelCommand { get; private set; }
+    public ICommand? ToggleConfigurationPanelCommand { get; private set; }
+    public ICommand? ToggleJobMenuPanelCommand { get; private set; }
+    public ICommand? ToggleFieldToolsPanelCommand { get; private set; }
+
+    // View Settings Commands
+    public ICommand? ToggleGridCommand { get; private set; }
+    public ICommand? ToggleDayNightCommand { get; private set; }
+    public ICommand? Toggle2D3DCommand { get; private set; }
+    public ICommand? ToggleNorthUpCommand { get; private set; }
+
+    // Camera Commands
+    public ICommand? IncreaseCameraPitchCommand { get; private set; }
+    public ICommand? DecreaseCameraPitchCommand { get; private set; }
+    public ICommand? IncreaseBrightnessCommand { get; private set; }
+    public ICommand? DecreaseBrightnessCommand { get; private set; }
+
+    // iOS Sheet Toggle Commands
+    public ICommand? ToggleFileMenuCommand { get; private set; }
+    public ICommand? ToggleFieldToolsCommand { get; private set; }
+    public ICommand? ToggleSettingsCommand { get; private set; }
+
+    // Map Commands
+    public ICommand? Toggle3DModeCommand { get; private set; }
+    public ICommand? ZoomInCommand { get; private set; }
+    public ICommand? ZoomOutCommand { get; private set; }
+
+    // Events for views to wire up to map controls
+    public event Action? ZoomInRequested;
+    public event Action? ZoomOutRequested;
+
+    private void InitializeNavigationCommands()
+    {
+        // Panel toggle commands
+        ToggleViewSettingsPanelCommand = new RelayCommand(() =>
+        {
+            IsViewSettingsPanelVisible = !IsViewSettingsPanelVisible;
+        });
+
+        ToggleFileMenuPanelCommand = new RelayCommand(() =>
+        {
+            IsFileMenuPanelVisible = !IsFileMenuPanelVisible;
+        });
+
+        ToggleToolsPanelCommand = new RelayCommand(() =>
+        {
+            IsToolsPanelVisible = !IsToolsPanelVisible;
+        });
+
+        ToggleConfigurationPanelCommand = new RelayCommand(() =>
+        {
+            IsConfigurationPanelVisible = !IsConfigurationPanelVisible;
+        });
+
+        ToggleJobMenuPanelCommand = new RelayCommand(() =>
+        {
+            IsJobMenuPanelVisible = !IsJobMenuPanelVisible;
+        });
+
+        ToggleFieldToolsPanelCommand = new RelayCommand(() =>
+        {
+            IsFieldToolsPanelVisible = !IsFieldToolsPanelVisible;
+        });
+
+        // View settings commands
+        ToggleGridCommand = new RelayCommand(() =>
+        {
+            IsGridOn = !IsGridOn;
+        });
+
+        ToggleDayNightCommand = new RelayCommand(() =>
+        {
+            IsDayMode = !IsDayMode;
+        });
+
+        Toggle2D3DCommand = new RelayCommand(() =>
+        {
+            Is2DMode = !Is2DMode;
+        });
+
+        ToggleNorthUpCommand = new RelayCommand(() =>
+        {
+            IsNorthUp = !IsNorthUp;
+        });
+
+        // Camera commands
+        IncreaseCameraPitchCommand = new RelayCommand(() =>
+        {
+            CameraPitch += 5.0;
+        });
+
+        DecreaseCameraPitchCommand = new RelayCommand(() =>
+        {
+            CameraPitch -= 5.0;
+        });
+
+        IncreaseBrightnessCommand = new RelayCommand(() =>
+        {
+            Brightness += 5; // Match the step from DisplaySettingsService
+        });
+
+        DecreaseBrightnessCommand = new RelayCommand(() =>
+        {
+            Brightness -= 5;
+        });
+
+        // iOS Sheet toggle commands
+        ToggleFileMenuCommand = new RelayCommand(() =>
+        {
+            IsFileMenuVisible = !IsFileMenuVisible;
+        });
+
+        ToggleFieldToolsCommand = new RelayCommand(() =>
+        {
+            IsFieldToolsVisible = !IsFieldToolsVisible;
+        });
+
+        ToggleSettingsCommand = new RelayCommand(() =>
+        {
+            IsSettingsVisible = !IsSettingsVisible;
+        });
+
+        // Map commands
+        Toggle3DModeCommand = new RelayCommand(() =>
+        {
+            _mapService.Toggle3DMode();
+            Is2DMode = !_mapService.Is3DMode;
+        });
+
+        ZoomInCommand = new RelayCommand(() =>
+        {
+            _mapService.Zoom(1.2);
+            ZoomInRequested?.Invoke();
+        });
+
+        ZoomOutCommand = new RelayCommand(() =>
+        {
+            _mapService.Zoom(0.8);
+            ZoomOutRequested?.Invoke();
+        });
+    }
+}
